refactor: proxy for logger data
The TCP server will now function as a proxy in between the logging stick
and the Ginlong monitoring portal. This allows the data to flow through the
system and not having to worry about sending the correct responses for each
message.

We still parse the messages and extract the measurement data to make it available
to other sinks.